# src/swagger.yml
basePath: /
consumes:
- application/json
definitions:
  Entry:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      currency:
        type: string
        x-go-name: Currency
      id:
        format: int32
        type: integer
        x-go-name: ID
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        format: int32
        type: integer
        x-go-name: UserID
    type: object
    x-go-package: github.com/bongster/golang_20210228/db/sqlc
  Transfer:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      from_user_id:
        format: int32
        type: integer
        x-go-name: FromUserID
      id:
        format: int32
        type: integer
        x-go-name: ID
      status:
        type: string
        x-go-name: Status
      to_user_id:
        format: int32
        type: integer
        x-go-name: ToUserID
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/bongster/golang_20210228/db/sqlc
  TransferTXResult:
    description: TransferTXResult is the result of the transfer transaction
    properties:
      from_user:
        $ref: '#/definitions/User'
      to_user:
        $ref: '#/definitions/User'
      transfer:
        $ref: '#/definitions/Transfer'
    type: object
    x-go-package: github.com/bongster/golang_20210228/db/sqlc
  User:
    properties:
      balance:
        format: int64
        type: integer
        x-go-name: Balance
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      currency:
        type: string
        x-go-name: Currency
      id:
        format: int32
        type: integer
        x-go-name: ID
      password:
        type: string
        x-go-name: Password
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/bongster/golang_20210228/db/sqlc
  createTransferRequest:
    description: createTransferRequest user request parameter
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      from_user_id:
        format: int32
        type: integer
        x-go-name: FromUserID
      to_user_id:
        format: int32
        type: integer
        x-go-name: ToUserID
    type: object
    x-go-package: github.com/bongster/golang_20210228/api
  createUserRequest:
    description: createUserRequest user request parameter
    properties:
      currency:
        type: string
        x-go-name: Currency
      password:
        type: string
        x-go-name: Password
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/bongster/golang_20210228/api
  loginUserRequest:
    description: loginUserRequest user login request parameter
    properties:
      password:
        type: string
        x-go-name: Password
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/bongster/golang_20210228/api
info:
  description: Documentation for Payment API
  title: classification Payment API.
  version: 1.0.0
paths:
  /entries:
    get:
      operationId: listEntry
      responses:
        "200":
          $ref: '#/responses/listEntyResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - entries
  /entries/{id}:
    get:
      operationId: getEntry
      parameters:
      - format: int32
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/getEntryResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - entries
  /login:
    get:
      operationId: loginUser
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/loginUserRequest'
      responses:
        "200":
          $ref: '#/responses/loginUserResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - users
  /register:
    post:
      operationId: createUser
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/createUserRequest'
      responses:
        "200":
          $ref: '#/responses/createUserResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - users
  /transfers:
    post:
      operationId: createTransfer
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/createTransferRequest'
      responses:
        "200":
          $ref: '#/responses/createTransferResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - transfers
  /users:
    get:
      operationId: listUser
      responses:
        "200":
          $ref: '#/responses/listUserResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - users
  /users/{id}:
    get:
      operationId: getUser
      parameters:
      - format: int32
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/getUserResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
      tags:
      - users
produces:
- application/json
responses:
  badRequestResponse:
    description: badRequestResponse response structure after invalid input from body
    schema:
      type: object
  createTransferResponse:
    description: createTransferResponse response structure after success create user
    schema:
      $ref: '#/definitions/TransferTXResult'
  createUserResponse:
    description: createUserResponse response structure after success create user
    schema:
      $ref: '#/definitions/User'
  genericError:
    description: ""
    schema:
      properties:
        Error:
          description: An optional field name to which this validation applies
          type: string
        Message:
          description: The validation message
          example: Expected type int
          type: string
      required:
      - Message
      type: object
  getEntryResponse:
    description: ""
    schema:
      $ref: '#/definitions/Entry'
  getUserResponse:
    description: ""
    schema:
      $ref: '#/definitions/User'
  listEntyResponse:
    description: listEntyResponse return response type from /entries API
    schema:
      items:
        $ref: '#/definitions/Entry'
      type: array
  listUserResponse:
    description: ""
    schema:
      items:
        $ref: '#/definitions/User'
      type: array
  loginUserResponse:
    description: ""
    headers:
      access_token:
        type: string
  someResponse:
    description: ""
    schema:
      properties:
        Message:
          description: The validation message
          example: Expected type int
          type: string
      required:
      - Message
      type: object
  validationError:
    description: ValidationError use this error when invalid data comming
    schema:
      properties:
        FieldName:
          description: An optional field name to which this validation applies
          type: string
        Message:
          description: The validation message
          example: Expected type int
          type: string
      required:
      - Message
      type: object
schemes:
- http
swagger: "2.0"